% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_funcs.R
\name{choose_palette}
\alias{choose_palette}
\title{Choose color palette}
\usage{
choose_palette(n, palettes = c("Dark2", "Set1", "Paired"), returnAll = F)
}
\arguments{
\item{n}{The number of colors needed.}

\item{palettes}{The choices of the palettes, must be available in
\link{RColorBrewer}.}

\item{returnAll}{If TRUE, all the eligible palletes are returned.
Otherwise only one is returned.}
}
\value{
A character vector of color palette names.
}
\description{
Given the number of colors needed, this function returns the name
of a color palette available in \link{RColorBrewer}. The returned
palette provides the most distinct colors as possible.
}
