% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_funcs.R
\name{get_available_palettes}
\alias{get_available_palettes}
\title{Get all the palette names}
\usage{
get_available_palettes(palSpace = "brewer")
}
\description{
Get all the palette names
}
\keyword{internal}
