% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_funcs.R
\name{mat_to_set}
\alias{mat_to_set}
\title{Convert matrix to list of sets}
\usage{
mat_to_set(mat, selected = NULL, presence = 1L, absence = 0L)
}
\arguments{
\item{mat}{Input matrix, can be \code{\link[=data.frame]{data.frame()}} or \linkS4class{data.table}}

\item{selected}{A vector of numbers to select which columns are to
be processed, default is \code{NULL}, i.e., all columns.}

\item{presence}{The values in each column indicating the presence of
a column in a row. Default is \code{1}, so a value of \code{1} in that
column will include that element in final set.}
}
\description{
This function converts a matrix (or data.frame, data.table)
to a list of vectors, which can be used as input for
\code{\link[VennDiagram:venn.diagram]{VennDiagram::venn.diagram()}}. The names of the list elements
are the column names of the input matrix.
}
