% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plots.R
\name{melt_mat}
\alias{melt_mat}
\title{Convert a matrix to long format}
\usage{
melt_mat(mat, rowIds = NULL, rowIdAsFactor = TRUE)
}
\arguments{
\item{rowIds}{The row ids.
\itemize{
\item NULL: row ids are constructed in the form of \verb{r<num>} where
\code{num} is the order of rows.
\item A length-1 vector of character or integer: this specifies the column
in the matrix which is used as row ids.
\item A character vector with length = \code{nrow(mat)}: this vector
provides row ids directly.
}}
}
\description{
This is achieved via \link{data.table}.
}
\keyword{internal}
