% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_funcs.R
\name{pick_colors}
\alias{pick_colors}
\title{Pick colors}
\usage{
pick_colors(n, palName = NA)
}
\arguments{
\item{n}{The number of colors to return.}

\item{palName}{A character string specifying a color palette
from R package \link{RColorBrewer}. Check
\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal.info} for available palettes. If
this parameter is \code{NA} (default), a palette will be chosen
from c("Set2", "Set1", "Set3"), depending on \code{n}. When n is
larger than the number of colors in the specified palette,
\code{\link[=colorRampPalette]{colorRampPalette()}} is used to interpolate the colors.}
}
\value{
A character vector of colors.
}
\description{
Given a number \code{n}, the function returns a color vector
of length \code{n}.
}
