% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plots.R
\name{plot_categorical_heatmap}
\alias{plot_categorical_heatmap}
\title{Plot heatmap using categorical data}
\usage{
plot_categorical_heatmap(
  mat,
  color = NULL,
  borderColor = "white",
  borderLwd = 0.3,
  title = NA,
  xlab = NA,
  ylab = NA,
  legendPos = "top",
  rowIds = NULL,
  ...
)
}
\arguments{
\item{mat}{The input matrix.}

\item{color}{Filling colors of the heatmap:
\itemize{
\item NULL: colors are picked using the function \code{\link[=pick_colors]{pick_colors()}}
\item named vector: element names are the values of the matrix and
element values are colors, such as c("0"="red","1"="green").
\item a vector of colors: similarly to using named vector, but colors
are assigned to matrix values in order.
\item a length-1 vector: this can be a palette name from
\link{RColorBrewer}, see \link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal.info} for values.
Then colors will be picked from the palette.
}}

\item{rowIds}{The row ids.
\itemize{
\item NULL: row ids are constructed in the form of \verb{r<num>} where
\code{num} is the order of rows.
\item A length-1 vector of character or integer: this specifies the column
in the matrix which is used as row ids.
\item A character vector with length = \code{nrow(mat)}: this vector
provides row ids directly.
}}
}
\description{
This function creates a heatmap from a matrix whose values are
categorical. It is based on \code{\link[ggplot2:geom_tile]{ggplot2::geom_tile()}}.
}
