% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plots.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Heatmap}
\usage{
plot_heatmap(mat, ...)
}
\arguments{
\item{mat}{A matrix, containing the values to be plotted.}
}
\description{
Create a heatmap based on the \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}} if the values
are numeric, or based on \code{\link[ggplot2:geom_tile]{ggplot2::geom_raster()}} if the values are
categorical (i.e., character).
}
