% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_plots.R
\name{plot_overlaps}
\alias{plot_overlaps}
\title{Plot of overlaps}
\usage{
plot_overlaps(dat = NULL, counts = NULL, ...)
}
\arguments{
\item{dat}{Input data, can be a \code{\link[=data.frame]{data.frame()}} or
\code{\link[=list]{list()}}, with each column or element being a
set of items, and the common items are counted
as overlap.}

\item{counts}{Counts of common and set-specific items.
Can be a named vector or list, each element being
the count of items in different categories (given
by the names).}
}
\value{
A \code{plot} object
}
\description{
This function creates plots to show the overlap
of the input data.
}
\details{
In default, it will generate
a Venn Diagram when the number of input sets is
smaller than 4, and generate an UpSet plot when
the set size is bigger.
}
