% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_plots.R
\name{venn_diagram}
\alias{venn_diagram}
\title{Plot venn diagram}
\usage{
venn_diagram(dat = NULL, color = NULL, ...)
}
\arguments{
\item{dat}{A list with each element being a set}

\item{color}{A length-2 character vector to provide
the colors for min and max values, forming a color
gradient between them. Default is \code{c("white","blue")}.}
}
\description{
Plot venn diagram
}
